Keepalive default true, mulighet for å sette interval selv og fikset keepalive funksjonaliteten som neppe har funket som den skal
